refactor(web): DRY up deploy flow, restore publish-status fallback, extract sub-components
- Create utils/publish-app.ts: shared publishApp() with best-effort
  status enrichment when publicUrl is missing from publish response
  (restores behavior dropped from lib/publish-api.ts)
- Move deploy-store.ts from domains/chat/ to stores/ (cross-domain
  shared state used by both chat-page and library-view)
- Consolidate library-view.tsx to use deploy store instead of 4 local
  useState hooks for deploy state (isDeploying, tokenDialog, etc.)
- Extract LibraryAppCard, LibraryDocumentCard, LibraryAppCardActionsMenu
  into dedicated component files (library-view 916→500 lines)
- Fix clearAppHtmlCache not called after app deletion (stale cache bug)
- DRY up publish result toast (5 identical blocks → showPublishResultToast)
