+++ apps/web/src/components/apps/library-date.ts
@@ -1,0 +1,8 @@
+export function formatLibraryDate(epochMs: number): string {
+  const date = new Date(epochMs);
+  return date.toLocaleDateString(undefined, {
+    day: "numeric",
+    month: "short",
+    year: date.getFullYear() !== new Date().getFullYear() ? "numeric" : undefined,
+  });
+}
