+++ apps/web/src/stores/deploy-store.ts
@@ -1,16 +1,13 @@
 /**
  * Zustand store for the app share/deploy lifecycle.
  *
- * Owns the in-flight UI state for two operations on the app viewer:
- * - **Share** — export the current app to a `.vellum` bundle.
- * - **Deploy** — publish the app to Vercel (with an intermediate token
+ * Owns the in-flight UI state for two operations:
+ * - **Share** — export an app to a `.vellum` bundle.
+ * - **Deploy** — publish an app to Vercel (with an intermediate token
  *   dialog when the org doesn't yet have a Vercel token stored).
  *
- * Split out from `useViewerStore` because none of these fields relate
- * to navigation (`mainView`, `intelligenceTab`) or viewer lifecycle
- * (`openedAppState`, `openedDocumentState`); they form an independent
- * data concern and only ship together with the app-share/deploy code
- * paths.
+ * Used by both the chat-page app viewer and the library page — lives
+ * in `stores/` because it is cross-domain shared state.
  *
  * Reference: {@link https://zustand.docs.pmnd.rs/}
  */
@@ -18,12 +15,11 @@
 import { create } from "zustand";
 
 import { toast } from "@vellum/design-library";
-import {
-  appsByIdPublishPost,
-  integrationsVercelConfigGet,
-} from "@/generated/daemon/sdk.gen";
+import { integrationsVercelConfigGet } from "@/generated/daemon/sdk.gen";
+import type { AppsByIdPublishPostResponse } from "@/generated/daemon/types.gen";
 import { isCredentialError } from "@/types/publish-types";
 import { createSelectors } from "@/utils/create-selectors";
+import { publishApp } from "@/utils/publish-app";
 import { shareApp as shareAppApi } from "@/utils/share-app";
 
 // ---------------------------------------------------------------------------
@@ -62,6 +58,24 @@ export interface DeployActions {
 }
 
 export type DeployStore = DeployState & DeployActions;
+
+// ---------------------------------------------------------------------------
+// Helpers
+// ---------------------------------------------------------------------------
+
+function showPublishResultToast(result: AppsByIdPublishPostResponse): void {
+  if (result.publicUrl) {
+    toast.success("Deployed to Vercel", {
+      description: result.publicUrl,
+      action: {
+        label: "Open",
+        onClick: () => window.open(result.publicUrl, "_blank"),
+      },
+    });
+  } else {
+    toast.success("Deployed to Vercel");
+  }
+}
 
 // ---------------------------------------------------------------------------
 // Initial state
@@ -137,26 +151,15 @@ const useDeployStoreBase = create<DeployStore>()((set, get) => ({
         set({ isTokenDialogOpen: true, pendingDeployAppId: appId, isDeploying: false });
         return;
       }
-      const { data: result } = await appsByIdPublishPost({
-        path: { assistant_id: assistantId, id: appId },
-        throwOnError: true,
-      });
+      const result = await publishApp(assistantId, appId);
       if (!result.success) {
         if (isCredentialError(result)) {
           set({ isTokenDialogOpen: true, pendingDeployAppId: appId, isDeploying: false });
         } else {
           toast.error("Failed to deploy", { description: result.error });
         }
-      } else if (result.publicUrl) {
-        toast.success("Deployed to Vercel", {
-          description: result.publicUrl,
-          action: {
-            label: "Open",
-            onClick: () => window.open(result.publicUrl, "_blank"),
-          },
-        });
       } else {
-        toast.success("Deployed to Vercel");
+        showPublishResultToast(result);
       }
     } catch (err) {
       toast.error("Failed to deploy", {
@@ -173,22 +176,11 @@ const useDeployStoreBase = create<DeployStore>()((set, get) => ({
     if (!pendingDeployAppId) return;
     set({ isDeploying: true });
     try {
-      const { data: result } = await appsByIdPublishPost({
-        path: { assistant_id: assistantId, id: pendingDeployAppId },
-        throwOnError: true,
-      });
+      const result = await publishApp(assistantId, pendingDeployAppId);
       if (!result.success) {
         toast.error("Failed to deploy", { description: result.error });
-      } else if (result.publicUrl) {
-        toast.success("Deployed to Vercel", {
-          description: result.publicUrl,
-          action: {
-            label: "Open",
-            onClick: () => window.open(result.publicUrl, "_blank"),
-          },
-        });
       } else {
-        toast.success("Deployed to Vercel");
+        showPublishResultToast(result);
       }
     } catch (err) {
       toast.error("Failed to deploy", {
